feat(condo): DOMA-6513 extracted LoginWithSBBOLButton to separate module and implemented new visual style
