+++ apps/condo/domains/common/components/containers/global-styles.css
@@ -213,3 +213,29 @@
 .ant-notification {
   z-index: 999;
 }
+
+/* Button */
+
+.condo-btn.condo-btn-sbbol {
+  border-color: var(--condo-global-color-teal-5);
+}
+
+.condo-btn.condo-btn-sbbol .condo-btn-text {
+  color: var(--condo-global-color-teal-5);
+}
+
+.condo-btn.condo-btn-sbbol:hover .condo-btn-text {
+  color: var(--condo-global-color-green-5);
+}
+
+.condo-btn.condo-btn-sbbol:active .condo-btn-text {
+  color: var(--condo-global-color-green-7);
+}
+
+.condo-btn.condo-btn-sbbol:hover .condo-btn-icon svg path {
+  fill: var(--condo-global-color-green-5);
+}
+
+.condo-btn.condo-btn-sbbol:active .condo-btn-icon svg path {
+  fill: var(--condo-global-color-green-7);
+}
